+++ src/Nethermind/Nethermind.Runner/ConsoleHelpers.cs
@@ -1,0 +1,43 @@
+// SPDX-FileCopyrightText: 2024 Demerzel Solutions Limited
+// SPDX-License-Identifier: LGPL-3.0-only
+
+using System;
+using System.Runtime.InteropServices;
+
+namespace Nethermind.Runner;
+
+public static class ConsoleHelpers
+{
+    public static void EnableConsoleColorOutput()
+    {
+        const int STD_OUTPUT_HANDLE = -11;
+        const uint ENABLE_VIRTUAL_TERMINAL_PROCESSING = 4;
+
+        Console.OutputEncoding = System.Text.Encoding.UTF8;
+
+        if (!OperatingSystem.IsWindowsVersionAtLeast(10))
+            return;
+
+        try
+        {
+            // If using Cmd and not set in registry
+            // enable ANSI escape sequences here
+            var handle = GetStdHandle(STD_OUTPUT_HANDLE);
+            GetConsoleMode(handle, out var mode);
+            mode |= ENABLE_VIRTUAL_TERMINAL_PROCESSING;
+            SetConsoleMode(handle, mode);
+        }
+        catch
+        {
+        }
+    }
+
+    [DllImport("kernel32.dll")]
+    private static extern bool GetConsoleMode(IntPtr hConsoleHandle, out uint lpMode);
+
+    [DllImport("kernel32.dll", SetLastError = true)]
+    private static extern IntPtr GetStdHandle(int nStdHandle);
+
+    [DllImport("kernel32.dll")]
+    private static extern bool SetConsoleMode(IntPtr hConsoleHandle, uint dwMode);
+}
